feat: UseCase 패키지 정리
