+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/domain/usecase/auth/SignInUseCase.kt
@@ -1,7 +1,7 @@
-package com.droidblossom.archive.domain.usecase
+package com.droidblossom.archive.domain.usecase.auth
 
 import android.util.Log
-import com.droidblossom.archive.data.dto.auth.request.VerificationNumberValidRequestDto
+import com.droidblossom.archive.data.dto.auth.request.SignInRequestDto
 import com.droidblossom.archive.domain.model.auth.Token
 import com.droidblossom.archive.domain.repository.AuthRepository
 import com.droidblossom.archive.util.RetrofitResult
@@ -12,13 +12,14 @@ import com.droidblossom.archive.util.onSuccess
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class ValidMessageUseCase @Inject constructor(
+class SignInUseCase @Inject constructor(
     private val repository: AuthRepository
 ) {
-    suspend operator fun invoke(request: VerificationNumberValidRequestDto) =
+
+    suspend operator fun invoke(request: SignInRequestDto) =
         flow<RetrofitResult<Token>> {
             try {
-                emit(repository.authValidMessage(request).onSuccess {
+                emit(repository.authSignIn(request).onSuccess {
 
                 }.onFail {
 
@@ -27,7 +28,7 @@ class ValidMessageUseCase @Inject constructor(
                 }.onException {
                     throw Exception(it)
                 })
-            } catch (e: Exception) {
+            } catch (e: Exception){
                 Log.d("예외확인", "$e")
                 e.printStackTrace()
             }
