+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/domain/usecase/auth/SignUpUseCase.kt
@@ -1,8 +1,8 @@
-package com.droidblossom.archive.domain.usecase
+package com.droidblossom.archive.domain.usecase.auth
 
 import android.util.Log
-import com.droidblossom.archive.data.dto.auth.request.SignInRequestDto
-import com.droidblossom.archive.domain.model.auth.Token
+import com.droidblossom.archive.data.dto.auth.request.SignUpRequestDto
+import com.droidblossom.archive.domain.model.auth.TemporaryToken
 import com.droidblossom.archive.domain.repository.AuthRepository
 import com.droidblossom.archive.util.RetrofitResult
 import com.droidblossom.archive.util.onError
@@ -12,23 +12,23 @@ import com.droidblossom.archive.util.onSuccess
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class SignInUseCase @Inject constructor(
+class SignUpUseCase @Inject constructor(
     private val repository: AuthRepository
 ) {
-
-    suspend operator fun invoke(request: SignInRequestDto) =
-        flow<RetrofitResult<Token>> {
+    suspend operator fun invoke(request: SignUpRequestDto) =
+        flow<RetrofitResult<TemporaryToken>> {
+            Log.d("후후후", "request ${request}")
             try {
-                emit(repository.authSignIn(request).onSuccess {
-
+                emit(repository.authSignUp(request).onSuccess {
+                    Log.d("후후후", "성공 - 재발급")
                 }.onFail {
-
+                    Log.d("후후후", "실후 - 재발급${it}")
                 }.onError {
                     throw Exception(it)
                 }.onException {
                     throw Exception(it)
                 })
-            } catch (e: Exception){
+            } catch (e: Exception) {
                 Log.d("예외확인", "$e")
                 e.printStackTrace()
             }
