+++ frontend/ARchive/app/src/main/java/com/droidblossom/archive/presentation/ui/auth/AuthViewModelImpl.kt
@@ -7,13 +7,13 @@ import com.droidblossom.archive.domain.model.auth.SignUp
 import com.droidblossom.archive.domain.model.auth.VerificationMessageSend
 import com.droidblossom.archive.domain.model.auth.VerificationNumberValid
 import com.droidblossom.archive.domain.model.member.CheckStatus
-import com.droidblossom.archive.domain.usecase.MemberStatusUseCase
-import com.droidblossom.archive.domain.usecase.ReIssueUseCase
-import com.droidblossom.archive.domain.usecase.SendMessageUseCase
-import com.droidblossom.archive.domain.usecase.SignInUseCase
-import com.droidblossom.archive.domain.usecase.SignUpUseCase
-import com.droidblossom.archive.domain.usecase.TemporaryTokenReIssueUseCase
-import com.droidblossom.archive.domain.usecase.ValidMessageUseCase
+import com.droidblossom.archive.domain.usecase.member.MemberStatusUseCase
+import com.droidblossom.archive.domain.usecase.auth.ReIssueUseCase
+import com.droidblossom.archive.domain.usecase.auth.SendMessageUseCase
+import com.droidblossom.archive.domain.usecase.auth.SignInUseCase
+import com.droidblossom.archive.domain.usecase.auth.SignUpUseCase
+import com.droidblossom.archive.domain.usecase.auth.TemporaryTokenReIssueUseCase
+import com.droidblossom.archive.domain.usecase.auth.ValidMessageUseCase
 import com.droidblossom.archive.presentation.base.BaseViewModel
 import com.droidblossom.archive.util.SharedPreferencesUtils
 import com.droidblossom.archive.util.onError
